refactor. 패키지명 수정
fix. hibernate 네이밍 설정 수정
