+++ src/main/java/io/gittul/GitTulBeApplication.java
@@ -1,4 +1,4 @@
-package io.gittul.gittulbe;
+package io.gittul;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
