+++ src/main/java/io/gittul/domain/github/entity/Developer.java
@@ -1,16 +1,19 @@
-package io.gittul.gittulbe.domain.github.entity;
+package io.gittul.domain.github.entity;
 
-import io.gittul.gittulbe.global.EntityTimeStamp;
+import io.gittul.global.jpa.EntityTimeStamp;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import lombok.AccessLevel;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
 
 @Entity
 @Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Developer extends EntityTimeStamp {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
