+++ src/main/java/io/gittul/domain/github/entity/GitHubRepository.java
@@ -1,16 +1,21 @@
-package io.gittul.gittulbe.domain.github.entity;
+package io.gittul.domain.github.entity;
 
-import io.gittul.gittulbe.global.EntityTimeStamp;
+import io.gittul.global.jpa.EntityTimeStamp;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
 
 @Entity
 @Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Table(name = "REPOSITORY")
 public class GitHubRepository extends EntityTimeStamp {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -25,6 +30,4 @@ public class GitHubRepository extends EntityTimeStamp {
     private Integer starCount;
 
     private Integer forkCount;
-
-    private LocalDateTime lastCrawledAt;
 }
