+++ src/main/java/io/gittul/domain/post/comment/entity/Comment.java
@@ -1,8 +1,8 @@
-package io.gittul.gittulbe.domain.post.comment.entity;
+package io.gittul.domain.post.comment.entity;
 
-import io.gittul.gittulbe.domain.post.entity.Post;
-import io.gittul.gittulbe.domain.user.entity.User;
-import io.gittul.gittulbe.global.EntityTimeStamp;
+import io.gittul.domain.post.entity.Post;
+import io.gittul.domain.user.entity.User;
+import io.gittul.global.jpa.EntityTimeStamp;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -10,13 +10,16 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
+import lombok.AccessLevel;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import java.util.ArrayList;
 import java.util.List;
 
 @Entity
 @Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Comment extends EntityTimeStamp {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
