+++ src/main/java/io/gittul/domain/post/entity/Post.java
@@ -1,10 +1,10 @@
-package io.gittul.gittulbe.domain.post.entity;
+package io.gittul.domain.post.entity;
 
-import io.gittul.gittulbe.domain.github.entity.GitHubRepository;
-import io.gittul.gittulbe.domain.post.comment.entity.Comment;
-import io.gittul.gittulbe.domain.user.entity.Bookmark;
-import io.gittul.gittulbe.domain.user.entity.User;
-import io.gittul.gittulbe.global.EntityTimeStamp;
+import io.gittul.domain.github.entity.GitHubRepository;
+import io.gittul.domain.post.comment.entity.Comment;
+import io.gittul.domain.user.entity.Bookmark;
+import io.gittul.domain.user.entity.User;
+import io.gittul.global.jpa.EntityTimeStamp;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -14,13 +14,16 @@ import jakarta.persistence.JoinTable;
 import jakarta.persistence.ManyToMany;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
+import lombok.AccessLevel;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import java.util.ArrayList;
 import java.util.List;
 
 @Entity
 @Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Post extends EntityTimeStamp {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
