+++ src/main/java/io/gittul/domain/post/entity/PostTag.java
@@ -1,6 +1,6 @@
-package io.gittul.gittulbe.domain.post.entity;
+package io.gittul.domain.post.entity;
 
-import io.gittul.gittulbe.global.EntityTimeStamp;
+import io.gittul.global.jpa.EntityTimeStamp;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
