+++ src/main/java/io/gittul/domain/post/entity/Tag.java
@@ -1,14 +1,17 @@
-package io.gittul.gittulbe.domain.post.entity;
+package io.gittul.domain.post.entity;
 
-import io.gittul.gittulbe.global.EntityTimeStamp;
+import io.gittul.global.jpa.EntityTimeStamp;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import lombok.AccessLevel;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Tag extends EntityTimeStamp {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
