+++ src/main/java/io/gittul/domain/post/entity/UserLikePost.java
@@ -1,24 +1,30 @@
-package io.gittul.gittulbe.domain.post.entity;
+package io.gittul.domain.post.entity;
 
-import io.gittul.gittulbe.domain.user.entity.User;
-import io.gittul.gittulbe.global.EntityTimeStamp;
+import io.gittul.domain.user.entity.User;
+import io.gittul.global.jpa.EntityTimeStamp;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import lombok.AccessLevel;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class UserLikePost extends EntityTimeStamp {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long likeId;
 
     @ManyToOne
+    @JoinColumn(name = "USER_ID")
     private User user;
 
     @ManyToOne
+    @JoinColumn(name = "POST_ID")
     private Post post;
 }
