+++ src/main/java/io/gittul/domain/user/UserRepository.java
@@ -1,0 +1,17 @@
+package io.gittul.domain.user;
+
+import io.gittul.domain.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByEmail(String email);
+
+    boolean existsByUserName(String s);
+
+    boolean existsByEmail(String email);
+}
