+++ src/main/java/io/gittul/domain/user/entity/Bookmark.java
@@ -1,19 +1,20 @@
-package io.gittul.gittulbe.domain.user.entity;
+package io.gittul.domain.user.entity;
 
-import io.gittul.gittulbe.domain.post.entity.Post;
-import io.gittul.gittulbe.global.EntityTimeStamp;
+import io.gittul.domain.post.entity.Post;
+import io.gittul.global.jpa.EntityTimeStamp;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import lombok.AccessLevel;
 import lombok.Getter;
-import lombok.Setter;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
-@Setter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Bookmark extends EntityTimeStamp {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
