+++ src/main/java/io/gittul/domain/user/entity/UserInterest.java
@@ -1,0 +1,30 @@
+package io.gittul.domain.user.entity;
+
+import io.gittul.domain.post.entity.Tag;
+import io.gittul.global.jpa.EntityTimeStamp;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class UserInterest extends EntityTimeStamp {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long interestId;
+
+    @ManyToOne
+    @JoinColumn(name = "USER_ID")
+    private User user;
+
+    @ManyToOne
+    @JoinColumn(name = "TAG_ID")
+    private Tag tag;
+}
