+++ src/test/java/io/gittul/GitTulBeApplicationTests.java
@@ -1,4 +1,4 @@
-package io.gittul.gittulbe;
+package io.gittul;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
